[Pet-to-MLIR] Pass type information to createOp
This commit is related to 08b9d2466f27c1a2f700a39cc47d1b5426656496.
Specifically, PET doesn't allow us to distinguish the width for a float
expression. This causes an issue if you need to create a pet_expr_double as we
don't know the width (i.e., f32 or f64). This commit passes the type to
createOp, which in turn may need the type to handle a pet_expr_double.